Updated Checkout, Admin Login and Product ID
- Fixed issue where the admin dashboard could open without logging in.
- Now user and admin info is fetched using their ID instead of email.
- User dashboard now shows order details, and product ID is saved in the orders table when an order is placed.

+++ Retail shop/admin/orders.php
@@ -17,7 +17,13 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['update_status'])) {
     exit;
 }
 ?>
-<?php require_once 'include/header.php'; ?>
+<?php
+require_once 'include/header.php';
+if (!isset($_SESSION['admin_id'])) {
+  echo "<script>window.open('login.php','_self');</script>";
+  exit();
+}
+?>
 <?php
     if (isset($_GET['delete_id'])) {
         $query1 = "DELETE FROM orders WHERE order_id = " . $_GET['delete_id'];
